day6/pyproject.toml
[project]
name = "day6"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv",
    "loguru",
]
